requirements.txt
# Core
python-dotenv>=1.0
pyyaml>=6.0

# Image Processing
opencv-python-headless>=4.8
numpy>=1.24
Pillow>=10.0

# OCR
pytesseract>=0.3.10
pdf2image>=1.16

# ML/Transformers
transformers>=4.36
torch>=2.1

# API
fastapi>=0.109
uvicorn>=0.25
python-multipart>=0.0.6

# Dev
pytest>=7.0
pytest-cov>=4.0
pytest-asyncio>=0.23
httpx>=0.26
ruff>=0.1
pre-commit>=3.0
